requirements.txt
pydobot==1.3.2
flask==3.0.2
tinydb==4.8.0
pyserial==3.4